djchat/requirements.txt
anyio==3.7.1
asgiref==3.7.2
attrs==23.1.0
channels==4.0.0
click==8.1.6
colorama==0.4.6
Django==4.2.4
django-cors-headers==4.2.0
django-dotenv==1.4.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-spectacular==0.26.4
exceptiongroup==1.1.3
h11==0.14.0
httptools==0.6.0
idna==3.4
inflection==0.5.1
jsonschema==4.18.6
jsonschema-specifications==2023.7.1
Pillow==10.0.0
PyJWT==2.8.0
python-dotenv==1.0.0
pytz==2023.3
PyYAML==6.0.1
referencing==0.30.2
rpds-py==0.9.2
sniffio==1.3.0
sqlparse==0.4.4
typing_extensions==4.7.1
tzdata==2023.3
uritemplate==4.1.1
uvicorn==0.23.2
watchfiles==0.19.0
websockets==11.0.3
